requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
PyMySQL==1.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
reportlab==4.0.7
python-dateutil==2.8.2


